Move action mapping demo to examples
